Refactoring of the file tabe into smaller files.
